Some fixes to windows dpi scaling

+++ indra/llwindow/llwindowwin32.cpp
@@ -2757,7 +2757,7 @@ LRESULT CALLBACK LLWindowWin32::mainWindowProc(HWND h_wnd, UINT u_msg, WPARAM w_
 		case WM_DPICHANGED:
 			{
 				LPRECT lprc_new_scale;
-				F32 new_scale = LOWORD(w_param) / USER_DEFAULT_SCREEN_DPI;
+				F32 new_scale = (float) LOWORD(w_param) / (float) USER_DEFAULT_SCREEN_DPI;
 				lprc_new_scale = (LPRECT)l_param;
 				S32 new_width = lprc_new_scale->right - lprc_new_scale->left;
 				S32 new_height = lprc_new_scale->bottom - lprc_new_scale->top;
@@ -4065,7 +4065,7 @@ F32 LLWindowWin32::getSystemUISize()
 				hr = pGDFM(hMonitor, MDT_EFFECTIVE_DPI, &dpix, &dpiy);
 				if (hr == S_OK)
 				{
-					scale_value = dpix / USER_DEFAULT_SCREEN_DPI;
+					scale_value = (float) dpix / (float) USER_DEFAULT_SCREEN_DPI;
 				}
 				else
 				{
@@ -4084,7 +4084,7 @@ F32 LLWindowWin32::getSystemUISize()
 	else
 	{
 		LL_WARNS() << "Could not load shcore.dll library (included by <ShellScalingAPI.h> from Win 8.1 SDK). Using legacy DPI awareness API of Win XP/7" << LL_ENDL;
-		scale_value = GetDeviceCaps(hdc, LOGPIXELSX) / USER_DEFAULT_SCREEN_DPI;
+		scale_value = (float) GetDeviceCaps(hdc, LOGPIXELSX) / (float) USER_DEFAULT_SCREEN_DPI;
 	}
 
 	ReleaseDC(hWnd, hdc);
